\name{cleaned}
\alias{cleaned}
\docType{data}
\title{
Daily filtered of suspect precipitation data example
}
\description{
Daily filtered of suspect precipitation data example. One month from '2001-01-01' to '2001-01-31' of 48 stations.
}
\usage{data("precipDataset")}
\format{
  A \code{matrix} with daily precipitation data without suspect values of five years from '2001-01-01' to '2001-01-31' of 48 stations.
}
\examples{
data(precipDataset)
}
\keyword{datasets}
