\name{filled}
\alias{filled}
\docType{data}
\title{
Daily reconstructed precipitation data example
}
\description{
Daily reconstructed precipitation data of one month from '2001-01-01' to '2001-01-31' of 48 stations.
}
\usage{data("precipDataset")}
\format{
  A \code{matrix} with daily reconstructed precipitation data of one month from '2001-01-01' to '2001-01-31' of 48 stations.
}
\examples{
data(precipDataset)
}
\keyword{datasets}
