% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqc_WeeklyCycle-get_wd_fraction.R
\name{get_wd_fraction}
\alias{get_wd_fraction}
\title{Weekly cycle test data frame}
\usage{
get_wd_fraction(xts_obj)
}
\arguments{
\item{xts_obj}{xts of a single time series}
}
\description{
The function get the results of the statistical test of the wet day fraction within the weekly cycle test
}
\references{
Hunziker, S., Gubler, S., Calle, J., Moreno, I., Andrade, M., Velarde, F., ... & Brönnimann, S. (2017). Identifying, attributing, and overcoming common data quality issues of manned station observations. https://doi.org/10.1002/joc.5037

Huerta, A., Serrano-Notivoli, R., & Brönnimann, S. (2024). SC-PREC4SA: A serially complete daily precipitation dataset for South America. https://doi.org/10.31223/X57D8R
}
