\name{preci}
\alias{preci}
\docType{data}
\title{
Daily precipitation data example
}
\description{
Daily precipitation data of five years from '2001-01-01' to '2005-12-31' of 48 stations.
}
\usage{data("precipDataset")}
\format{
  A \code{matrix} with daily precipitation data of five years from '2001-01-01' to '2005-12-31' of 48 stations.
}
\examples{
data(precipDataset)
}
\keyword{datasets}
